Move ProjectDocument type
