+++ src/Types.ts
@@ -131,23 +131,6 @@ export interface Document {
     meta?: object;
   };
 }
-
-export interface ProjectDocument {
-  id: string;
-
-  created_at: string;
-
-  created_by: string;
-
-  updated_at?: string;
-
-  updated_by?: string;
-
-  project_id: string;
-
-  document_id: string;
-}
-
 export interface DocumentInContext extends Document {
   layers: Layer[];
 }
